project/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
Django==3.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
pika==1.2.0
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.1.1
wincertstore==0.2
